+++ src/metatrain/utils/metadata.py
@@ -1,5 +1,6 @@
 import collections.abc
 import json
+from typing import Optional
 
 from metatomic.torch import ModelMetadata
 
@@ -20,12 +21,17 @@ def update(d, u):
     return d
 
 
-def merge_metadata(self: ModelMetadata, other: ModelMetadata) -> ModelMetadata:
+def merge_metadata(
+    self: ModelMetadata, other: Optional[ModelMetadata] = None
+) -> ModelMetadata:
     """Append ``references`` to an existing ModelMetadata object.
 
     :param self: The metadata object to be updated.
-    :param other: The metadata object to merged to self.
+    :param other: The metadata object to merged to self. If None, return self.
     """
+
+    if other is None:
+        return self
 
     self_dict = json.loads(self._get_method("__getstate__")())
     other_dict = json.loads(other._get_method("__getstate__")())